requirements.txt
pillow==12.2.0
rembg==2.0.75
streamlit==1.56.0
